Extract code from ArchiveSerialization.h. Optimize header dependencies.
