+++ Source/Urho3D/IO/ArchiveSerializationContainer.h
@@ -1,0 +1,301 @@
+//
+// Copyright (c) 2017-2020 the rbfx project.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+//
+
+#pragma once
+
+#include "../IO/ArchiveSerializationBasic.h"
+#include "../IO/ArchiveSerializationVariant.h"
+
+namespace Urho3D
+{
+
+namespace Detail
+{
+
+/// Serialize tie of vectors of the same size. Each tie of elements is serialized as separate block.
+template <class T, class U, size_t... Is>
+inline void SerializeVectorTie(Archive& archive, const char* name, const char* element, T& vectorTuple, const U& serializeValue, ea::index_sequence<Is...>)
+{
+    const unsigned sizes[] = { ea::get<Is>(vectorTuple).size()... };
+
+    unsigned numElements = sizes[0];
+    auto block = archive.OpenArrayBlock(name, numElements);
+
+    if (archive.IsInput())
+    {
+        numElements = block.GetSizeHint();
+        (ea::get<Is>(vectorTuple).clear(), ...);
+        (ea::get<Is>(vectorTuple).resize(numElements), ...);
+    }
+    else
+    {
+        if (ea::find(ea::begin(sizes), ea::end(sizes), numElements, ea::not_equal_to<unsigned>{}) != ea::end(sizes))
+            throw ArchiveException("Vectors of '{}/{}' have mismatching sizes", archive.GetCurrentBlockPath(), name);
+    }
+
+    for (unsigned i = 0; i < numElements; ++i)
+    {
+        const auto elementTuple = ea::tie(ea::get<Is>(vectorTuple)[i]...);
+        serializeValue(archive, element, elementTuple);
+    }
+}
+
+}
+
+/// Serialize vector with standard interface. Content is serialized as separate objects.
+template <class T, class TSerializer = Detail::DefaultSerializer>
+void SerializeVectorAsObjects(Archive& archive, const char* name, const char* element, T& vector, const TSerializer& serializeValue = TSerializer{})
+{
+    using ValueType = typename T::value_type;
+
+    unsigned numElements = vector.size();
+    auto block = archive.OpenArrayBlock(name, numElements);
+
+    if (archive.IsInput())
+    {
+        numElements = block.GetSizeHint();
+        vector.clear();
+        vector.resize(numElements);
+    }
+
+    for (unsigned i = 0; i < numElements; ++i)
+        serializeValue(archive, element, vector[i]);
+}
+
+/// Serialize array with standard interface (compatible with ea::span, ea::array, etc). Content is serialized as separate objects.
+template <class T, class TSerializer = Detail::DefaultSerializer>
+void SerializeArrayAsObjects(Archive& archive, const char* name, const char* element, T& array, const TSerializer& serializeValue = TSerializer{})
+{
+    const unsigned numElements = array.size();
+    auto block = archive.OpenArrayBlock(name, numElements);
+
+    if (archive.IsInput())
+    {
+        if (numElements != block.GetSizeHint())
+            throw ArchiveException("'{}/{}' has unexpected array size", archive.GetCurrentBlockPath(), name);
+    }
+
+    for (unsigned i = 0; i < numElements; ++i)
+        serializeValue(archive, element, array[i]);
+}
+
+template <class T, class TSerializer = Detail::DefaultSerializer>
+void SerializeVectorTieAsObjects(Archive& archive, const char* name, const char* element, T vectorTuple, const TSerializer& serializeValue = TSerializer{})
+{
+    static constexpr auto tupleSize = ea::tuple_size_v<T>;
+    Detail::SerializeVectorTie(archive, name, element, vectorTuple, serializeValue, ea::make_index_sequence<tupleSize>{});
+}
+
+/// Serialize vector with standard interface. Content is serialized as bytes.
+template <class T>
+void SerializeVectorAsBytes(Archive& archive, const char* name, T& vector)
+{
+    using ValueType = typename T::value_type;
+    static_assert(std::is_standard_layout<ValueType>::value, "Type should have standard layout to safely use byte serialization");
+    static_assert(std::is_trivially_copyable<ValueType>::value, "Type should be trivially copyable to safely use byte serialization");
+
+    const bool loading = archive.IsInput();
+    ArchiveBlock block = archive.OpenUnorderedBlock(name);
+
+    unsigned sizeInBytes{};
+
+    if (!loading)
+        sizeInBytes = vector.size() * sizeof(ValueType);
+
+    archive.SerializeVLE("size", sizeInBytes);
+
+    if (loading)
+    {
+        if (sizeInBytes % sizeof(ValueType) != 0)
+            throw ArchiveException("'{}/{}' has unexpected size in bytes", archive.GetCurrentBlockPath(), name);
+        vector.resize(sizeInBytes / sizeof(ValueType));
+    }
+
+    archive.SerializeBytes("data", vector.data(), sizeInBytes);
+}
+
+/// Serialize vector in the best possible format.
+template <class T>
+void SerializeVector(Archive& archive, const char* name, const char* element, T& vector)
+{
+    using ValueType = typename T::value_type;
+    static constexpr bool standardLayout = std::is_standard_layout<ValueType>::value;
+    static constexpr bool triviallyCopyable = std::is_trivially_copyable<ValueType>::value;
+
+    if constexpr (standardLayout && triviallyCopyable)
+    {
+        if (!archive.IsHumanReadable())
+        {
+            SerializeVectorAsBytes(archive, name, vector);
+            return;
+        }
+    }
+
+    SerializeVectorAsObjects(archive, name, element, vector);
+}
+
+/// Serialize custom vector.
+/// While writing, serializer may skip vector elements. Size should match actual number of elements to be written.
+/// While reading, serializer must push elements into vector on its own.
+template <class T, class U>
+void SerializeCustomVector(Archive& archive, const char* name, unsigned sizeToWrite, const T& vector, U serializer)
+{
+    using ValueType = typename T::value_type;
+    auto block = archive.OpenArrayBlock(name, sizeToWrite);
+    if (archive.IsInput())
+    {
+        for (unsigned index = 0; index < block.GetSizeHint(); ++index)
+        {
+            static const ValueType valuePlaceholder{};
+            serializer(index, valuePlaceholder, true);
+        }
+    }
+    else
+    {
+        for (unsigned index = 0; index < vector.size(); ++index)
+            serializer(index, vector[index], false);
+    }
+}
+
+/// Serialize map or hash map with with standard interface.
+template <class T, class TSerializer = Detail::DefaultSerializer>
+void SerializeMap(Archive& archive, const char* name, const char* element, T& map, const TSerializer& serializeValue = TSerializer{})
+{
+    using KeyType = typename T::key_type;
+    using ValueType = typename T::mapped_type;
+
+    if constexpr (ea::is_same_v<ValueType, Variant> && ea::is_same_v<TSerializer, Detail::DefaultSerializer>)
+    {
+        auto block = archive.OpenArrayBlock(name, map.size());
+        if (archive.IsInput())
+        {
+            map.clear();
+            for (unsigned i = 0; i < block.GetSizeHint(); ++i)
+            {
+                auto elementBlock = archive.OpenUnorderedBlock(element);
+                KeyType key{};
+                SerializeValue(archive, "key", key);
+                SerializeVariantInBlock(archive, elementBlock, map[key]);
+            }
+        }
+        else
+        {
+            for (const auto& [key, value] : map)
+            {
+                auto elementBlock = archive.OpenUnorderedBlock(element);
+                SerializeValue(archive, "key", const_cast<KeyType&>(key));
+                SerializeVariantInBlock(archive, elementBlock, const_cast<Variant&>(value));
+            }
+        }
+    }
+    else
+    {
+        auto block = archive.OpenArrayBlock(name, map.size());
+        if (archive.IsInput())
+        {
+            map.clear();
+            for (unsigned i = 0; i < block.GetSizeHint(); ++i)
+            {
+                auto elementBlock = archive.OpenUnorderedBlock(element);
+                KeyType key{};
+                serializeValue(archive, "key", key);
+                serializeValue(archive, "value", map[key]);
+            }
+        }
+        else
+        {
+            for (const auto& [key, value] : map)
+            {
+                auto elementBlock = archive.OpenUnorderedBlock(element);
+                serializeValue(archive, "key", const_cast<KeyType&>(key));
+                serializeValue(archive, "value", const_cast<ValueType&>(value));
+            }
+        }
+    }
+}
+
+/// Serialize set or hash set with standard interface.
+template <class T, class TSerializer = Detail::DefaultSerializer>
+void SerializeSet(Archive& archive, const char* name, const char* element, T& set, const TSerializer& serializeValue = TSerializer{})
+{
+    using ValueType = typename T::value_type;
+
+    auto block = archive.OpenArrayBlock(name, set.size());
+    if (archive.IsInput())
+    {
+        set.clear();
+        for (unsigned i = 0; i < block.GetSizeHint(); ++i)
+        {
+            auto elementBlock = archive.OpenUnorderedBlock(element);
+            ValueType value{};
+            serializeValue(archive, element, value);
+            set.emplace(ea::move(value));
+        }
+    }
+    else
+    {
+        for (const auto& value : set)
+            serializeValue(archive, element, const_cast<ValueType&>(value));
+    }
+}
+
+/// Serialize shared pointer to Object.
+template <class T, std::enable_if_t<std::is_base_of<Object, T>::value, int> = 0>
+void SerializeValue(Archive& archive, const char* name, SharedPtr<T>& value)
+{
+    const bool loading = archive.IsInput();
+    ArchiveBlock block = archive.OpenUnorderedBlock(name);
+
+    StringHash type{};
+    ea::string_view typeName{};
+    if (!loading && value)
+    {
+        type = value->GetType();
+        typeName = value->GetTypeName();
+    }
+
+    SerializeStringHash(archive, "type", type, typeName);
+
+    if (loading)
+    {
+        // Serialize null object
+        if (type == StringHash{})
+        {
+            value = nullptr;
+            return;
+        }
+
+        // Create instance
+        Context* context = archive.GetContext();
+        value.StaticCast(context->CreateObject(type));
+        if (!value)
+        {
+            throw ArchiveException("Failed to create object '{}/{}' of type {}", archive.GetCurrentBlockPath(), name,
+                type.ToDebugString());
+        }
+    }
+
+    if (value)
+        SerializeValue(archive, "value", *value);
+}
+
+}
